requirements.txt
anthropic==0.3.6
Pillow==9.5.0
pdf2image==1.16.3
colorama==0.4.6
